refactor : 타 도메인 에러 코드 수정
- Member, Pay 도메인의 임포트 에러 코드를 수정했습니다.
