+++ src/main/java/com/dangdang/server/domain/memberTown/application/MemberTownService.java
@@ -5,11 +5,10 @@ import com.dangdang.server.domain.member.domain.MemberRepository;
 import com.dangdang.server.domain.member.domain.entity.Member;
 import com.dangdang.server.domain.memberTown.domain.MemberTownRepository;
 import com.dangdang.server.domain.memberTown.domain.entity.MemberTown;
-import com.dangdang.server.domain.memberTown.domain.entity.RangeType;
 import com.dangdang.server.domain.memberTown.dto.request.MemberTownSaveRequest;
 import com.dangdang.server.domain.memberTown.dto.response.MemberTownSaveResponse;
+import com.dangdang.server.domain.town.domain.TownRepository;
 import com.dangdang.server.domain.town.domain.entity.Town;
-import com.dangdang.server.domain.town.domain.entity.TownRepository;
 import com.dangdang.server.global.exception.BusinessException;
 import com.dangdang.server.global.exception.ExceptionCode;
 import java.util.List;
